qcacmn: Clear the radar flag when NOL count is zero
Consider STA vap to be up in chan 100 VHT80 mode and stadfs enabled.
Say Radar is injected on chan 100 and 100 is added to NOL list and
IEEE80211_CHAN_SET_RADAR is set on it. On NOL timer expiry, during the
process of nol delete, dfs nol count is decremented. So it reads 0 now.
After this decrement, dfs_nol_update() is called and it returns without
calling dfs_mlme_clist_update as nol count is zero.
As dfs_mlme_clist_update is not invoked, IEEE80211_CHAN_CLR_RADAR flag is
not set. After NOL timeout, STA vap fails to associate with Root on VHT80
mode as 100 VHT80 is not available (chan 100’s RADAR flag not cleared).

CRs-Fixed: 2327546
Change-Id: Ic83ec2ac49e0fb7e887c3c7be2f8eefc419cb7c5