qcacmn: Remove load/unload log in diag reporting
There are multiple places in the driver where the diag events
are reported and if the unload has started these diag events
are logged and dropped if there is storm of diag events because
of some thread message flushing this logging of the message will
cause excessive logging.

Remove the log as driver is unloading and it won't be sending it
userspace logging either.

Change-Id: Ie4c5caf44ad509da213f1bc76ff6ef6dfeecde2d
CRs-Fixed: 2716899