qcacmn: Check for valid vdev queue for serialization
While removing command from serialization, it is possible
that with the last command removal of a vdev,
the vdev could be deleted and before attempting the
movement of command from pending vdev queue to active queue,
we need check if the serialization vdev object is valid.

Add a check to ensure that serialization vdev object is
valid before trying to move command from pending to active
for a vdev.

CRs-Fixed: 2807736
Change-Id: Iea6df72eeeb62b7868f30337e7da99039df8b23a