qcacmn: Change default scan max timeout value for CM scan
The current value of scan timeout value is 10 secs.
Earlier for ap platform the scan timeout wass set to default as 50 secs
and and also for ap platform the scan could be triggered for internally
triggered connect and the max scan timeout
of 50 secs should not result in supplicant timeouts.

So setting the max scan timeout value of 50 sec for ap platforms

CRs-Fixed: 2870782
Change-Id: Icd2841f1f5503623f15bdbdea3430ef6559b9a1b