qcacmn: Disable DFS for 5/10MHZ channels
Add an API reg_disable_dfs_for_half_qtr_chans() which disables DFS
flag in regulatory current channel list if reg chan's max BW is less
than 20MHZ.
If radar is found on a 20MHZ BW, to choose the next random channel,
utils_dfs_get_chan_list() is invoked. It returns a list of all supported
5G channels of the pdev which includes half/qtr channels as well.
If the next chosen channel has a max BW of 5/10MHZ and the current chanbw
is full rate, we will not be able to find a channel pointer and come up.
Hence channel change after radar fails.
Hence ensure that only full rate supported channels are retrieved by
utils_dfs_get_chan_list() API.

reg_is_dfs_for_freq:
If the DUT is operating on 20MHZ BW or greater (and not 5/10MHZ),
evaluate the regulatory channel flag of the given freq and determine
if it is DFS or not.
If the DUT is operating on 5/10MHZ BW (read from user config chanbw),
the given input channel is considered as non-DFS irrespective of
the regulatory channel flag configuration.

CRs-Fixed: 3044601
Change-Id: I485f031211bbfa6c905af47d9e889386533a0c4d