qcacmn: Populate half/qtr flag in radiotap header
In the radiotap header of the wireless frame, there are bit fields
to indicate if the packet is received on a half/quarter rate channel.
This bit must be populated by the ucode and sent to the host via
PPDU TLV so that it can parsed and set in the radio tap header.

Currently these bits are not populated by the ucode. As a workaround,
these bit fields are set based on the monitor vap's current channel
(ic->curchan) flag. In mlme_ext_vap_up_pre_init(), the current
channel flag is set in dp_pdev via CDP_MONITOR_CHAN_FLAG pdev param.
Based on the value of this flag, pdev->ppdu_info.rx_status is populated
to set half/quarter flags and the corresponding radtio tap header
bits are set in qdf_nbuf_update_radiotap().

CRs-Fixed: 3053723
Change-Id: I8d27078d43cd166af1bad985c0c7939944132aed