qcacmn: Modify chan_to_freq for 2.5MHZ step chans
When 2.5MHZ step size channels are enabled in the master channel
list and channel 2 (2417 MHZ) is given as the input to
reg_compute_chan_to_freq() API, the API computes the frequency
incorrectly.

The 2.5MHZ step channel numbers are not sorted in ascending order
in 2.4GHZ regulatory channel map (eg: IEEE 1, 203, 2, 204 etc).
Hence when reg_compute_chan_to_freq() does a linear search in the
master channel list, the following condition satisfies
"chan_list[count].chan_num >= chan_num" and hence it breaks out of the
search in master channel list. It then applies a default channel to freq
formula which is incorrect for a 2.5MHZ step chan.

Hence handle 2.5MHZ step chans as an exception. Break out
of the search of the master channel list, only if the input channel number
matches with the master channel list chan enum.

CRs-Fixed: 3079547
Change-Id: I7cbfbcbab9e231637c3db74c17c721d0b3b3e228