qcacmn: Update NOL cleared status for individual 20MHz channels
NOL entries are maintained as ranges. Hence, if NOL expires for
a single range of 80MHz, the preCAC unmarking API is only called
for one channel in that range, even if there are 4 20MHz channels
which expired.

Add an API to find the 20MHz channels that are expired and call
the action to update preCAC entries and notify the upper layer for
all of the individual 20MHz entries.

Change-Id: Id360cc3058dd8276802de6467bc4d1f9a44271f8
CRs-Fixed: 3136930