qcacmn: Validate vdev count before pdev CSA switch count update
The num of vdevs parameter of the pdev csa switch count TLV is a
tainted value and is not checked before use. Invalid values can
cause a slab out of bound error, which results in a crash.

Add a check for the num of vdevs while extracting pdev csa
switch count TLV. Compare the number of vdevs with the num of
vdev ids parameter of the csa switch count firmware event. Drop
the event if the check fails.

Change-Id: I88401984437186a1a6e077c274d0011fa93e8704
CRs-Fixed: 4503619